refactor : charge test code move to pathacceptancetest
